Bring back the auditor; completion of initial unit tests
There aren't many unit tests for the grabber; I'm sure more will be
created over time, but the bulk of the testing for the grabber really
must be done with integration testing against the SD service.  Most of
what unit testing would catch is covered by the Auditor -- a form of
FVT.  Though designed to identify issues upstream at Schedules Direct,
the Auditor class also will catch grabber bugs with the generation of
the cache file.  So the testing pattern for the grabber should be to run
the (limited) unit tests, run SD integration tests (to be written
still), then run the Auditor against a recent cache file.  Once that all
passes, we're good to go.