Address lost days in local zipped schedules
Since the beginning there has been a bug in this grabber that would
cause days to be lost in schedules stored in the zip cache file.
Basically, the grabber logic to pull down only days that changed worked
except that when those changed days were written to the zip it would
first wipe the schedule file from the zip before rewriting it.  The
problem is it was only rewriting the days that changed (since that's all
that was downloaded from the server).  The end result is that days in
the cache that hadn't changed were then lost.  Since SageTV already had
those days in its db, I never noticed this bug.  Thanks to a new
(non-Sage) user, this bug was brought to my attention and now fixed.